styling and cleanup of "my positions" table components
