+++ app/components/decisionTables/DecisionResultCell.js
@@ -1,8 +1,13 @@
 import React from 'react'
+import { Text } from '@aragon/ui'
 import MarketNameStyled from './MarketNameStyled'
 
 const DecisionResultCell = ({ decisionResult }) => (
-  <MarketNameStyled type={decisionResult} />
+  <div>
+    <Text size="xsmall">
+      <MarketNameStyled type={decisionResult} />
+    </Text>
+  </div>
 )
 
 export default DecisionResultCell
